Refactored socket events of node into separate file
